webapp/requirements.txt
requests
pandas
numpy
beautifulsoup
flask
redis
psycopg2-binary